Multiple suggestions on NonResidenceExtractID
The current version only provides non-residence events for individuals showing residence events in at least two distinct receivers, therefore excluding individuals that only visit a single acoustic station. Suggestion #1 avoids excluding these individuals and allows to establish non-residence events for all the individuals that show more than 1 residence event. 
The current version provides 0s values for DISTANCE and ROM when the distance matrix is not available. Suggestion #2 allows to replace these 0s by NAs, which is more appropriate for unknown information.
The current version provides a ROM of 0 for non-residence events from and to the same receiver (recursion). While the DISTANCE is known to be 0 (which is also debatable), the ROM remains unknown. Suggestion #3 allows to have NAs inseatd of 0s in this case.
Suggestion #4 only allows to get rid of the IF instruction at the end of the function and provides nothing more than the reduction of the length of the code.